requirements.txt
PySide6>=6.10,<7
Pillow>=10,<12
